+++ pkg/commands/service/version/serviceversion_test.go
@@ -323,6 +323,26 @@ func TestVersionStage(t *testing.T) {
 			WantError: "error parsing arguments: required flag --version not provided",
 		},
 		{
+			Name: "validate --autoclone stages a clone of an active version",
+			Args: "--service-id 123 --version 1 --autoclone",
+			API: &mock.API{
+				GetVersionFn:      testutil.GetVersion,
+				CloneVersionFn:    testutil.CloneVersionResult(4),
+				ActivateVersionFn: stageVersionOK,
+			},
+			WantOutput: "Staged service 123 version 4",
+		},
+		{
+			Name: "validate --autoclone stages a clone of a locked version",
+			Args: "--service-id 123 --version 2 --autoclone",
+			API: &mock.API{
+				GetVersionFn:      testutil.GetVersion,
+				CloneVersionFn:    testutil.CloneVersionResult(4),
+				ActivateVersionFn: stageVersionOK,
+			},
+			WantOutput: "Staged service 123 version 4",
+		},
+		{
 			Args: "--service-id 123 --version 3",
 			API: &mock.API{
 				GetVersionFn:      testutil.GetVersion,